♻️패키지 구조 수정
엔티티들은 domain -> entity 패키지로 이동하여 dto와 분리하였습니다.
