+++ src/main/java/dnd/donworry/domain/entity/OptionImage.java
@@ -1,0 +1,21 @@
+package dnd.donworry.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class OptionImage {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @OneToOne
+    private Selection selection;
+
+    @Column(nullable = false)
+    private String path;
+}
