+++ src/main/java/dnd/donworry/domain/entity/UserVote.java
@@ -1,4 +1,4 @@
-package dnd.donworry.domain;
+package dnd.donworry.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -8,18 +8,16 @@ import lombok.*;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class VoteLike extends BaseEntity {
+public class UserVote extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
     @ManyToOne
-    private Vote vote;
-
-    @ManyToOne
     private User user;
 
-    @Column(nullable = false)
-    private boolean status = true;
+    @ManyToOne
+    private Selection selection;
+
 }
