+++ src/main/java/dnd/donworry/domain/entity/Vote.java
@@ -1,9 +1,7 @@
-package dnd.donworry.domain;
+package dnd.donworry.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.List;
 
 @Entity
 @Getter
